# [NPU] refine update_loss_scaling npu kernel