<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs

### Describe
<!-- Describe what this PR does -->

use `ZerosLike` and `Memcpy` instead of `NPUMemsetAsync`.

- before (use NPUMemsetAsync)
<img width="1115" alt="图片" src="https://user-images.githubusercontent.com/26408901/116392062-8fe70a00-a852-11eb-90c4-0ed6dccc7350.png">

As shown in the timeline, there is a blank correspondding to `update_loss_scaling_op` caused by `NPUMemsetAsync`.
`update_loss_scaling_op` cost about 103 ms.

![图片](https://user-images.githubusercontent.com/26408901/116395093-40a2d880-a856-11eb-8fec-5162ef73de0b.png)


- only use ZerosLike
If only use `ZerosLike` to replace `NPUMemsetAsync`.
<img width="1084" alt="图片" src="https://user-images.githubusercontent.com/26408901/116393270-020c1e80-a854-11eb-8c78-8bf65ab0a4c3.png">

`update_loss_scaling_op` will launch many `ZerosLike` NPU ops.
`update_loss_scaling_op` cost about 22.2 ms.

- In this PR, use ZerosLike and Memcpy
<img width="1092" alt="图片" src="https://user-images.githubusercontent.com/26408901/116393331-16e8b200-a854-11eb-943d-ba8b8fdc5edc.png">

`update_loss_scaling_op` will launch only 1 `ZerosLike` NPU op, and then use `Memcpy` to set tensor to 0.
`update_loss_scaling_op` cost about 5.5 ms.


![图片](https://user-images.githubusercontent.com/26408901/117416811-55fdbe00-af4c-11eb-9670-85270f39072d.png)



### Performance

Speed up: 19448 tokens/s -> 20679 tokens/s, +6.33  %